tmq client router operators
